notodications and compliance update

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -15,7 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('email')->unique();
-            $table->integer('phone')->nullable();
+            $table->bigInteger('phone')->nullable();
+
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->boolean('status')->default(1);
